+++ src/main/java/proyecto/integrador/clinica/service/impl/PacienteService.java
@@ -10,6 +10,7 @@ import proyecto.integrador.clinica.dto.response.DomicilioResponseDto;
 import proyecto.integrador.clinica.dto.response.PacienteResponseDto;
 import proyecto.integrador.clinica.entity.Domicilio;
 import proyecto.integrador.clinica.entity.Paciente;
+import proyecto.integrador.clinica.exception.ResourceNotFoundException;
 import proyecto.integrador.clinica.repository.IPacienteRepository;
 import proyecto.integrador.clinica.service.IPacienteService;
 
@@ -42,7 +43,12 @@ public class PacienteService implements IPacienteService {
 
     @Override
     public Optional<Paciente> buscarPorId(Integer id) {
-        return pacienteRepository.findById(id);
+        Optional<Paciente> pacienteEncontrado = pacienteRepository.findById(id);
+        if(pacienteEncontrado.isPresent()){
+            return pacienteEncontrado;
+        } else {
+            throw new ResourceNotFoundException("Paciente no encontrado");
+        }
     }
 
     @Override
@@ -69,7 +75,13 @@ public class PacienteService implements IPacienteService {
 
     @Override
     public void eliminarPaciente(Integer id) {
-        pacienteRepository.deleteById(id);
+        Optional<Paciente> pacienteEncontrado = buscarPorId(id);
+        if (pacienteEncontrado.isPresent()){
+            pacienteRepository.deleteById(id);
+        }else{
+            throw new ResourceNotFoundException("Paciente no encontrado");
+        }
+
 
     }
